pyproject.toml
[project]
name = "foo-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.7",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "langgraph-cli[inmem]>=0.4.7",
    "python-dotenv>=1.2.1",
]
